Make it possible to override the issuer
This makes configuring the issuer more flexible.

For example, when using a framework like Symfony, it is preferable to
not use the global `$_SERVER` array, but to use the `Request` object to
get the host.

There might also be other cases where someone does not want to use the
domainname from the `Host` HTTP header, but some other canonical
domainname.